Fix a2x option collection from input file with non-ascii encoding (#104)
* Fix a2x Option Collection From Input File With Non-ASCII Encoding

As an alternative to passing all of the options on the command line, a2x allows one to specify options in the body of the text of the source file (e.g. lines starting with "//a2x: <option>" will have "<option>" processed as if it was on the command line).  Even if a user does not use this feature, a2x will always perform the scan.

The code that scans the input file for options does not specify an encoding.  So, in those cases where the actual encoding of the file does not match the locale (e.g. input file is UTF-8, yet user has set LC_ALL="C" in the environment), there is a UnicodeDecodeError exception thrown.

The fix is to open the file in binary mode when scanning for options.  The line will only be attempted to be converted to ASCII if it already matches the regular expression for an a2x option.  If the line cannot be converted to ASCII, then a warning message will be issued, and execution will continue.

NOTE:  This change requires that all a2x options specified on the command line be normal ASCII text (which is most likely everyone's desire anyway).

Co-authored-by: Matthew Peveler <matt.peveler@gmail.com>