remove A-A-P build files
This is a defunct build system that not even its creator seems to use.
While they have not been totally reverse engineered, it does not seem
like anyone has used these in years, and no one has complained about it,
so probably safe to remove and re-implement actual requested functionality
into Makefile as it comes up.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>