remove slash after $(DESTDIR) usage
The slash between $(DESTDIR) and the target was causing an extra
slash as generally the target already started with a slash, leading
to something like //usr/local/etc/. The double slash would be ignored
on Linux, but treated as a network path on Windows. The GNU make docs
also indicate not having the slash is probably best practices. See
https://www.gnu.org/prep/standards/html_node/DESTDIR.html.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>