fix handling of options attribute for DelimintedBlock (#120)
When attempting to do something like:

["test",options="textual"]
--------------------
Hello
--------------------

The options would not be parsed per the spec (https://asciidoc.org/userguide.html#X75) where you would not get textual-option as one of the attributes. The cause of this is 0b7a7d4 where AttributeList.consume was no longer being called with an argument we were using to hold all our attributes. Within this function is the special handling of the options attribute into parts that was no longer being called.

This re-adds special handling to get the expected <__>-option attributes out of options while retaining the behavior introduced in that patch. This is somewhat similar to how the Table works (in that it has its own inline parsing of options to have it work).

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>