Use config newline setting in system attribute evaluation (#155)
Problems seen with inconsistent newlines in the Table Of Contents HTML pulled in from the configuration file.  While the rest of the contents in the resulting HTML file honored the newline style configured by the user, this specific snippet of HTML always used "\r\n".

This root of this problem existed for some time, but in earlier versions, the newline style was always "\n".  The symptoms changed as a side effect of the issue "Extra line padding in source and literal blocks" (#139).  In this issue, the newline style changed to "\r\n", and was noticed by users.

Change the System Attribute Evaluation function to use the newline setting from the global "config" instead of using the default newline.