Allow python to be provided in Makefile (#276)
When 'python3' is used to do the install (through `make install`)
then the shbang on the installed program will be /usr/bin/python3.
For systems with more than one python, it is better to invoke
with python3.X so that the shbang has the proper value
(/usr/bin/python3.X rather than /usr/bin/python3).